// src/types/Customer.ts
export type Customer = {
  id: number;
  fullname: string;
  birthdate: string;
  gender: boolean;
  phonenumber: string;
  provinceId: number;
  province: string;
  areaId: number;
  area: string;
  educationId: number;
  education: string;
  workId: number;
  work: string;
  hobbyId: number;
  hobby: string;
};